requirements.txt
fastapi[all]==0.95.2
uvicorn[standard]==0.22.0
scikit-learn==1.3.2
librosa==0.10.0.post2
numpy==1.26.5
scipy==1.11.3
pydub==0.25.1
soundfile==0.12.1
python-dotenv==1.0.1
joblib==1.3.2
pytest==7.4.2
httpx==0.24.1
python-multipart==0.0.6